Rename classpath runner package
Remove `runner` since we're no longer tied to JUnit 4.

See gh-17491
